Readme
Very unhappy with this project, because I tried to make it something I'd want to see online, and I'd want to be proud to show my friends.  I'm not done with it by any means. But I have to walk away from it for my mind's sake.  I'm getting used to submitting things before they are due for a "paper trail of progress", and I had something I could have submitted yesterday and the day before, but I felt like I was just spinning wheels and couldn't get things to work.  Finally figured some of them out.  Not defeated by any means, just need to keep pushing and figure out where the blocks are.